// licentaApp/src/app/models/contact.ts
export interface Contact {
  firstName: string;
  lastName: string;
  email: string;
  phoneNumber: string;
  message: string;
  medicProblem: boolean;
  appointmentProblem: boolean;
  scheduleProblem: boolean;
  accountProblem: boolean;
  otherProblem: boolean;
  medicJoin: boolean;
  otherReason: boolean;
}